Move AwsRequestSigningApacheInterceptor to aws-plugin-api, use in os source and sink
